+++ server/src/migrations/V67__more_armor_migrations.sql
@@ -1,0 +1,24 @@
+UPDATE item
+    SET item_definition_id = 'common.items.armor.adventurer.belt'
+    WHERE item_definition_id = 'common.items.armor.twigs.belt';
+UPDATE item
+    SET item_definition_id = 'common.items.armor.adventurer.chest'
+    WHERE item_definition_id = 'common.items.armor.twigs.chest';
+UPDATE item
+    SET item_definition_id = 'common.items.armor.adventurer.foot'
+    WHERE item_definition_id = 'common.items.armor.twigs.foot';
+UPDATE item
+    SET item_definition_id = 'common.items.armor.adventurer.hand'
+    WHERE item_definition_id = 'common.items.armor.twigs.hand';
+UPDATE item
+    SET item_definition_id = 'common.items.armor.adventurer.pants'
+    WHERE item_definition_id = 'common.items.armor.twigs.pants';
+UPDATE item
+    SET item_definition_id = 'common.items.armor.adventurer.shoulder'
+    WHERE item_definition_id = 'common.items.armor.twigs.shoulder';
+UPDATE item
+    SET item_definition_id = 'common.items.armor.buccaneer.head'
+    WHERE item_definition_id = 'common.items.armor.pirate.hat';
+UPDATE item
+    SET item_definition_id = 'common.items.utility.lockpick.lockpick_copper'
+    WHERE item_definition_id = 'common.items.utility.lockpick_0';